IoC over files filtering
