+++ src/filters/index.js
@@ -1,5 +1,7 @@
 import DefaultFilter from './default.js'
+import { DefaultFilterLogic } from './defaultLogic'
 
 export {
   DefaultFilter,
+  DefaultFilterLogic,
 }
